refactor: Separate strategy out of database.py + put runners in their own dir.
